Added script docker socket vulnerbility
