+++ docs/api/running_container.md
@@ -191,4 +191,4 @@ api/docker/running_containers
     }
 }
 ```json
-````
+
